Cleanup dice rolling
Added some comments and changed how the doce rolling sends the results
back to the user in a way that is hopefully useable for all future
functions as well. The discord message commands can now accept any kind
of string, to make it easier to send whaatever you've got.

The other big change is that all handle_commands should now return a
future containing the message they want to send back to the user. In
theory all commands will want to send *something* back, and this way of
doing it ensures it can be batched along with the omnidata save so that
both can complete as fast as possible, rather than needing to wait until
after the save is complete. This should make for a much snappier bot.